@@ -9,37 +9,38 @@
   "author": "Christophe Duchesne-Ashworth <christophe.ashworth@gmail.com>",
   "license": "MIT",
   "private": false,
+  "type": "module",
   "files": [
     "/lib"
   ],
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/plugin-syntax-jsx": "^7.16.0",
-    "@babel/preset-typescript": "^7.16.0",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.0.6",
-    "@storybook/addon-actions": "^6.3.12",
-    "@storybook/addon-essentials": "^6.3.12",
-    "@storybook/addon-links": "^6.3.12",
-    "@storybook/react": "^6.3.12",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "babel-loader": "^8.2.3",
-    "eslint": "^8.2.0",
-    "eslint-plugin-react": "^7.27.1",
-    "postcss": "^8.3.11",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "rollup": "^2.58.3",
+    "@babel/core": "^7.21.3",
+    "@babel/plugin-syntax-jsx": "^7.18.6",
+    "@babel/preset-typescript": "^7.21.0",
+    "@rollup/plugin-commonjs": "^24.0.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@storybook/addon-actions": "^6.5.16",
+    "@storybook/addon-essentials": "^6.5.16",
+    "@storybook/addon-links": "^6.5.16",
+    "@storybook/react": "^6.5.16",
+    "@types/react": "^18.0.31",
+    "@types/react-dom": "^18.0.11",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
+    "babel-loader": "^9.1.2",
+    "eslint": "^8.37.0",
+    "eslint-plugin-react": "^7.32.2",
+    "postcss": "^8.4.21",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rollup": "^3.20.2",
     "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-postcss": "^4.0.1",
-    "rollup-plugin-scss": "3",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "sass": "^1.43.4",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.4.4"
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-scss": "4",
+    "rollup-plugin-typescript2": "^0.34.1",
+    "sass": "^1.60.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.2"
   },
   "peerDependencies": {
     "react": "^17.0.2",
@@ -55,10 +56,10 @@
     "lint:fix": "yarn eslint src/**/*.ts src/**/*.tsx --fix"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@types/jest": "^27.0.2",
-    "jest": "^27.3.1",
-    "ts-jest": "^27.0.7"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@types/jest": "^29.5.0",
+    "jest": "^29.5.0",
+    "ts-jest": "^29.0.5"
   }
 }
